requirements.txt
fastapi>=0.111,<0.112
uvicorn[standard]>=0.29,<0.30
SQLAlchemy>=2.0,<2.1
sqlmodel>=0.0.14,<0.0.15
aiomysql>=0.2,<0.3
PyMySQL>=1.1,<2.0
pydantic-settings>=2.2,<2.3
passlib[bcrypt]>=1.7,<1.8
bcrypt>=3.2,<4.0
PyJWT>=2.8,<3.0
python-multipart>=0.0.9
email-validator>=2.1,<3.0
